+++ src/components.d.ts
@@ -977,6 +977,10 @@ export namespace Components {
          */
         "active": boolean;
         /**
+          * When true, the tab will be rendered with a close icon.
+         */
+        "closable": boolean;
+        /**
           * Set to true to draw the tab in a disabled state.
          */
         "disabled": boolean;
@@ -2504,9 +2508,17 @@ declare namespace LocalJSX {
          */
         "active"?: boolean;
         /**
+          * When true, the tab will be rendered with a close icon.
+         */
+        "closable"?: boolean;
+        /**
           * Set to true to draw the tab in a disabled state.
          */
         "disabled"?: boolean;
+        /**
+          * Emitted when the tab is closable and the close button is activated.
+         */
+        "onSl-close"?: (event: CustomEvent<any>) => void;
         /**
           * The name of the tab panel the tab will control. The panel must be located in the same tab group.
          */