+++ docs/getting-started/changelog.md
@@ -14,7 +14,8 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Fixed a bug where `sl-menu-item` wouldn't render properly in the dark theme
 - Fixed a bug where `sl-select` would show an autocomplete menu
 - Improved placeholder contrast in dark theme
-- Updated Boostrap Icons to 1.1.0
+- Updated to Boostrap Icons 1.1.0
+- Updated to Stencil 2.3.0
 
 ## 2.0.0-beta.22